move utility functions to new file
